+++ spec/requests/v1/groups_spec.rb
@@ -179,9 +179,9 @@ RSpec.describe "V1::Groups", type: :request do
       let!(:group_slots) {
         create_list(:group_slot, 3, :with_media, group: group) }
 
-      it "deletes all photos on the group slots" do
-        g_slots_img_first = group.group_slots.first.photos.first
-        g_slots_img_last = group.group_slots.first.photos.last
+      it "deletes all images on the group slots" do
+        g_slots_img_first = group.group_slots.first.images.first
+        g_slots_img_last = group.group_slots.first.images.last
         delete "/v1/groups/#{group.id}", {}, auth_header
         g_slots_img_first.reload
         g_slots_img_last.reload